Extract WorkerExecutorFixture to enable reuse
